Move HTTP proxy code to a public package
